requirements.txt
discord.py==2.5.2
aiohttp==3.11.14
async-timeout==5.0.1
pycparser==2.22
cffi==1.17.1
imageio==2.37.0
pyNaCl==1.5.0
python-dotenv==1.1.0
yt-dlp==2023.3.4
ffmpeg==1.4
ffmpeg-python==0.2.0
psutil==7.0.0
multidict==6.2.0
numpy==2.2.4
attrs==25.3.0
aiohappyeyeballs==2.6.1
frozenlist==1.5.0
idna==3.10
yarl==1.18.3
opuslib==3.0.1
pillow==11.1.0
propcache==0.3.1
